backend/pyproject.toml
[project]
name = "ag-tools-catalogue-backend"
version = "1.0.0"
description = "Backend API for AG Tools Catalogue"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "httpx>=0.25.2",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "chromadb>=0.4.22",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "httpx>=0.25.2",
]
